Changed Java build process to use Apache Maven.
