+++ examples/plotsim.py
@@ -69,51 +69,51 @@ def main(args):
 
     plt.figure(0)
     plt.subplot(611)
-    plt.plot(tim, sma, "-b")
+    plt.plot(tim, sma, "ob")
     plt.xlabel("Time [hr]")
     plt.ylabel("SMA [km]")
     plt.subplot(612)
-    plt.plot(tim, ecc, "-b")
+    plt.plot(tim, ecc, "ob")
     plt.xlabel("Time [hr]")
     plt.ylabel("Eccentricity")
     plt.subplot(613)
-    plt.plot(tim, inc, "-b")
+    plt.plot(tim, inc, "ob")
     plt.xlabel("Time [hr]")
     plt.ylabel("Inclination [deg]")
     plt.subplot(614)
-    plt.plot(tim, raan, "-b")
+    plt.plot(tim, raan, "ob")
     plt.xlabel("Time [hr]")
     plt.ylabel("RAAN [deg]")
     plt.subplot(615)
-    plt.plot(tim, argp, "-b")
+    plt.plot(tim, argp, "ob")
     plt.xlabel("Time [hr]")
     plt.ylabel("Perigee arg. [deg]")
     plt.subplot(616)
-    plt.plot(tim, tran, "-b")
+    plt.plot(tim, tran, "ob")
     plt.xlabel("Time [hr]")
     plt.ylabel("True anom. [deg]")
 
     plt.figure(1)
     plt.subplot(211)
-    plt.plot(tim, hmag, "-b")
+    plt.plot(tim, hmag, "ob")
     plt.xlabel("Time [hr]")
     plt.ylabel("Angular momentum")
     plt.subplot(212)
-    plt.plot(tim, ener, "-b")
+    plt.plot(tim, ener, "ob")
     plt.xlabel("Time [hr]")
     plt.ylabel("Specific energy")
 
     plt.figure(2)
     plt.subplot(311)
-    plt.plot(tim, hvec[:,0], "-b")
+    plt.plot(tim, hvec[:,0], "ob")
     plt.xlabel("Time [hr]")
     plt.ylabel("h(x)")
     plt.subplot(312)
-    plt.plot(tim, hvec[:,1], "-b")
+    plt.plot(tim, hvec[:,1], "ob")
     plt.xlabel("Time [hr]")
     plt.ylabel("h(y)")
     plt.subplot(313)
-    plt.plot(tim, hvec[:,2], "-b")
+    plt.plot(tim, hvec[:,2], "ob")
     plt.xlabel("Time [hr]")
     plt.ylabel("h(z)")
     plt.suptitle("Components of angular momentum")
