--- orbdetpy/lib/orekit/NOTICE.txt
@@ -1,44 +1,0 @@
-OREKIT
-Copyright 2002-2018 CS Systèmes d'Information
-
-This product includes software developed by
-CS Systèmes d'Information (http://www.c-s.fr/)
-
-This product includes software developed by
-Bruce R. Bowman (HQ AFSPC, Space Analysis Division)
-
-This product includes software translated from original work developed by
-David A. Vallado, Paul Crawford, Richard Hujsak, T.S. Kelso
-
-This product includes software translated from original work developed by
-Felix R. Hoots, Ronald L. Roehrich
-
-This product includes software translated from original work developed by
-R. Biancale, S. Bruinsma (CNES)
-
-This product includes data computed by
-P. Gégout (CNRS / UMR5563 - GET)
-
-This product includes software translated from original work developed by
-R. Biancale, S. Bruinsma (CNES)
-
-This product includes software translated from original work developed by
-Mickaël Gastineau (CNRS - Observatoire de Paris - IMCCE)
-
-This product includes software translated from original work developed by
-The Apache Software Foundation (http://www.apache.org/)
-
-This product depends on software developed by
-The Apache Software Foundation (http://www.apache.org/)
-
-This product depends on software developed by
-The Hipparchus project (https://hipparchus.org/)
-
-This product includes software translated from original work developed by
-M. Picone, A.E. Hedin, D. Drob (Naval Research Laboratory)
-
-This product includes software translated from original work developed by
-Dominik Brodowski
-
-This product includes software translated from original work developed by
- Duncan Agnew and copyright 2008 IERS Conventions center
