+++ setup.py
@@ -54,7 +54,7 @@ PROJECT_URLS = {"Documentation": "https://github.com/ut-astria/orbdetpy",
                 "Source": "https://github.com/ut-astria/orbdetpy",
                 "Tracker": "https://github.com/ut-astria/orbdetpy/issues"}
 PACKAGE_DATA = {"orbdetpy": ["data/*", "data/**/*", "data/**/**/*",
-                             "lib/*", "lib/**/*"]}
+                             "target/astria*.jar"]}
 
 setup(name="orbdetpy",
       packages=["orbdetpy"],
